LayoutTests/http/tests/performance/performance-resource-timing-entries-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS performance.getEntriesByType('resource').length is 2
